feat(services): decimal discount percent (NUMERIC(7,3)) (#37)
Replaces INT discount_percent with NUMERIC(7,3), enabling admins to
set fractional discounts (e.g., 50.003%) needed for cost-based pricing
(e.g., cost \$300 -> sell at \$149.99 requires ~50.003% off).

Schema:
- Migration 0016: ALTER COLUMN INT -> NUMERIC(7,3) with USING cast
- Existing INT values (e.g., 10) cleanly become 10.000 - no data loss
- models.py: Mapped[Decimal | None] + Numeric(7,3)
- schemas.py: Decimal with max_digits=7, decimal_places=3
- Validation: strict bounds 0 < x < 100 (was 1 <= x <= 100)
- field_serializer for JSON number output (matches discount_amount pattern)

Order snapshot:
- discount_percent stored as str(Decimal) in service_snapshot JSONB for
  JSON portability (matches discount_amount_* pattern)

Pricing logic - UNCHANGED:
- calculate_discounted_price already uses Decimal arithmetic
- \$64.99 * (100 - 10.000) / 100 = \$58.49 (identical to INT path)

Tests: 6 new (4 unit + 2 integration) + 325 full suite green
OpenAPI snapshot: type=integer -> type=number on 4 surfaces

+++ app/features/services/models.py
@@ -102,7 +102,10 @@ class ServiceOption(Base, TimestampMixin):
     # rule. Order-level discounts (e.g. USDT 5%) stack on top of any
     # item-level discount because the order subtotal is computed from the
     # already-discounted unit price.
-    discount_percent: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # NUMERIC(7,3) lets the admin target sub-percent rates (12.5%, 7.499%).
+    # Integers stored as 10 (legacy) read back as Decimal("10.000") — the
+    # downstream Decimal math produces identical totals after quantize.
+    discount_percent: Mapped[Decimal | None] = mapped_column(Numeric(7, 3), nullable=True)
     discount_amount_usd: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
     discount_amount_eur: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
